requirements.txt
discord.py>=2.3.0
motor>=3.1.0
aiohttp>=3.8.0
Pillow>=9.0.0
python-dotenv>=0.19.0
requests>=2.28.0
reportlab>=4.0.0
flask
PyNaCl
yt-dlp
google-genai
reportlab
sift-stack-py
python-dotenv
psutil
